backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.95.0",
    "asyncpg>=0.31.0",
    "exa-py>=2.11.0",
    "fastapi>=0.135.3",
    "langchain-openai>=1.1.0",
    "langchain-xai>=1.2.2",
    "langgraph>=1.1.6",
    "matplotlib>=3.10.8",
    "openai>=2.31.0",
    "pgvector>=0.4.2",
    "pyjwt[crypto]>=2.12.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "scipy>=1.17.1",
    "sentence-transformers>=5.4.1",
    "uvicorn>=0.44.0",
]
